# specs/008-authenticated-backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Job Application Workspace Online API
  version: 0.1.0
  description: Authenticated online workspace and complete replacement operations.
servers:
  - url: https://app.example.com
security:
  - sessionCookie: []
  - desktopBearer: [workspace]
paths:
  /api/workspace:
    get:
      operationId: getWorkspace
      responses:
        '200':
          description: Current account workspace and revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/workspace/revision:
    get:
      operationId: getWorkspaceRevision
      responses:
        '200':
          description: Lightweight connectivity and revision check.
          content:
            application/json:
              schema:
                type: object
                required: [revision, updatedAt]
                properties:
                  revision: { type: integer, minimum: 0 }
                  updatedAt: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/positions:
    get:
      operationId: listPositions
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: Account-owned position summaries and workspace revision.
          content:
            application/json:
              schema:
                type: object
                required: [positions, listView, revision]
                properties:
                  positions:
                    type: array
                    items: { $ref: '#/components/schemas/PositionSummary' }
                  listView: { $ref: '#/components/schemas/ListView' }
                  revision: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createPosition
      parameters:
        - $ref: '#/components/parameters/ExpectedWorkspaceRevision'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PositionInput' }
      responses:
        '201':
          description: Created position and new workspace revision.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PositionMutationResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/RevisionConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/positions/{positionId}:
    parameters:
      - $ref: '#/components/parameters/PositionId'
    get:
      operationId: getPosition
      responses:
        '200':
          description: Complete account-owned position.
          content:
            application/json:
              schema:
                type: object
                required: [position, workspaceRevision]
                properties:
                  position: { $ref: '#/components/schemas/Position' }
                  workspaceRevision: { type: integer, minimum: 0 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updatePosition
      parameters:
        - $ref: '#/components/parameters/ExpectedWorkspaceRevision'
        - $ref: '#/components/parameters/ExpectedRecordRevision'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PositionInput' }
      responses:
        '200':
          description: Updated position and revisions.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PositionMutationResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/RevisionConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/positions/order:
    patch:
      operationId: reorderPositions
      parameters:
        - $ref: '#/components/parameters/ExpectedWorkspaceRevision'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderedIds]
              properties:
                orderedIds:
                  type: array
                  minItems: 1
                  uniqueItems: true
                  items: { type: string }
      responses:
        '200':
          description: New order and workspace revision.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/RevisionConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/positions/list-view:
    patch:
      operationId: updateListView
      parameters:
        - $ref: '#/components/parameters/ExpectedWorkspaceRevision'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ListView' }
      responses:
        '200': { description: Updated preference and workspace revision. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/RevisionConflict' }
  /api/positions/{positionId}/questions:
    parameters:
      - $ref: '#/components/parameters/PositionId'
    post:
      operationId: createQuestion
      parameters:
        - $ref: '#/components/parameters/ExpectedWorkspaceRevision'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '201': { description: Created question and updated revisions. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/RevisionConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/positions/{positionId}/readiness/readings:
    parameters:
      - $ref: '#/components/parameters/PositionId'
    post:
      operationId: createReading
      parameters:
        - $ref: '#/components/parameters/ExpectedWorkspaceRevision'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '201': { description: Created reading and updated revisions. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/RevisionConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    desktopBearer:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://app.example.com/api/auth/oauth2/authorize
          tokenUrl: https://app.example.com/api/auth/oauth2/token
          scopes:
            workspace: Access the owner's workspace.
  parameters:
    PositionId:
      in: path
      name: positionId
      required: true
      schema: { type: string }
    ExpectedWorkspaceRevision:
      in: header
      name: If-Match
      required: true
      schema: { type: string, description: Quoted workspace revision ETag. }
    ExpectedRecordRevision:
      in: header
      name: X-Record-Revision
      required: false
      schema: { type: integer, minimum: 0 }
  responses:
    Unauthorized:
      description: Missing, expired, revoked, or invalid authentication.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource absent or not owned by the authenticated account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RevisionConflict:
      description: Expected revision is stale; reload or create a new synchronization preview.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request or workspace package failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            issues:
              type: array
              items:
                type: object
                required: [path, message]
                properties:
                  path: { type: string }
                  message: { type: string }
    WorkspaceResponse:
      type: object
      required: [workspace, revision, updatedAt]
      properties:
        workspace:
          type: object
          description: Document conforming to the shared current PositionsDocument and ReferenceData schemas.
          additionalProperties: true
        revision: { type: integer, minimum: 0 }
        updatedAt: { type: string, format: date-time }
    PositionSummary:
      type: object
      required: [id, company, title, status, workMode, seniority, updatedAt]
      properties:
        id: { type: string }
        company: { type: object, additionalProperties: true }
        title: { type: string }
        status: { type: string }
        workMode: { type: string }
        seniority: { type: string }
        updatedAt: { type: string, format: date-time }
    Position:
      allOf:
        - $ref: '#/components/schemas/PositionInput'
        - type: object
          required: [id, revision, createdAt, updatedAt]
          properties:
            id: { type: string }
            revision: { type: integer, minimum: 0 }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    PositionInput:
      type: object
      description: Conforms to the shared current CreatePositionInput or PositionDetailsUpdate schema.
      additionalProperties: true
    PositionMutationResponse:
      type: object
      required: [position, workspaceRevision]
      properties:
        position: { $ref: '#/components/schemas/Position' }
        workspaceRevision: { type: integer, minimum: 0 }
    ListView:
      type: object
      required: [mode]
      properties:
        mode: { type: string, enum: [manual, column] }
        column: { type: [string, 'null'] }
        direction: { type: [string, 'null'], enum: [asc, desc, null] }
...